+++ apinext/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "apinext",
+  "name": "@beep/apinext",
   "version": "0.0.1",
   "description": "tRPC for Ride Beep App",
   "main": "src/index.ts",
@@ -20,7 +20,6 @@
     "@trpc/server": "11.0.0-rc.477",
     "cors": "^2.8.5",
     "drizzle-orm": "^0.32.2",
-    "ioredis": "^5.4.1",
     "postgres": "^3.4.4",
     "redis": "^4.7.0",
     "wkx": "^0.5.0"
